Update generated beta Java models and requests with build 4215428

+++ src/main/java/com/microsoft/graph/models/extensions/IpReferenceData.java
@@ -37,7 +37,7 @@ public class IpReferenceData implements IJsonBackedObject {
      */
     @SerializedName(value = "asn", alternate = {"Asn"})
     @Expose
-    public Integer asn;
+    public Long asn;
 
     /**
      * The City.
